backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-text-splitters
langgraph
langchain-qdrant
qdrant-client
pydantic
python-dotenv
python-multipart
pypdf
docx2txt
markdown
mlflow
httpx
langchain-google-genai
